### 2021-06-01 #3244 @arfon
:wave: @faroit – how about this as a second paper to edit? Given your expertise in web technologies it feels like it could be a good fit?
