### 2021-06-06 #3244 @faroit
👋  Hi @alicia-ziying-yang, thanks for your submission to JOSS! I will be serving as the editor for this. We'll use this issue to identify reviewers and resolve technical issues. Since @sara-02 has already volunteered (👏) for review we are looking for only one additional reviewer. Based on the list of people that have agreed to review for JOSS I found the following three potential reviewers to be a good fit:

- @ baileythegreen
- @ jaeyk
- @ cmaimone

@alicia-ziying-yang, I am also interested in your input on these reviewers, so any suggestions you have are also welcome, especially w.r.t. the JOSS [COI policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy). Please only mention potential reviewers with a space between the @ and their handle (so you would refer to me as "@ faroit").
