+++ doc/paper/paper.bib
@@ -94,7 +94,8 @@
   year="1994",
   publisher="Springer London",
   pages="292--300",
-  isbn="978-1-4471-2099-5"
+  isbn="978-1-4471-2099-5",
+  url="https://doi.org/10.1007/978-1-4471-2099-5_30"
 }
 
 @article{lda,
