### 2021-06-09 #3244 @faroit
@alicia-ziying-yang

> Hi, I have fixed the suggested DOI. Please check again.

thanks, looks good.

Can you also modify the markdown section headers so that they match the [JOSS template](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography)? E.g. `# 2. Statement of Need` -> `# Statement of need`... etc.
