+++ app.py
@@ -7,6 +7,7 @@ import dash_uploader as du
 import uuid
 import dash_bootstrap_components as dbc
 from dash.dependencies import Input, Output, State,MATCH, ALL
+from dash.exceptions import PreventUpdate
 import dash_table
 import plotly.graph_objs as go
 import dash_daq as daq
@@ -95,6 +96,7 @@ def build_banner(corpus_element,saves_element,loaded):
                     html.Label("From"),
                     daq.NumericInput(
                         id="year-from",
+                        style={"color":"#e85e56"},
                         className="year-selector",
                         min=min(year_list),
                         max=max(year_list),
@@ -102,6 +104,7 @@ def build_banner(corpus_element,saves_element,loaded):
                     ),html.Label("To"),
                     daq.NumericInput(
                         id="year-to",
+                        style={"color":"#e85e56"},
                         className="year-selector",
                         min=min(year_list),
                         max=max(year_list),
@@ -534,16 +537,43 @@ def build_tabs():
         ],
     )
 
-def build_tab_1():
+# --- build "sentences" tab ---
+# search, and show ranking
+
+def build_tab_1(corpus_name):
+    corpus_ind_dir=corpus[corpus_name]
+    display_fileds = preprocess_corpus.get_table_fieldnames(corpus_ind_dir)
+    display_fileds.append("score")
+    opt=[{'label': f.upper(), 'value': f} for f in display_fileds]
+    
     return [
         dcc.Store(id='memory-output'),
         dcc.Loading(color="rgba(240, 218, 209,0.8)",type="cube",style={"padding-top":"25%"},children=[
+            html.Div(style={"margin-left": "30px","margin-top": "1.5%", "font-variant": "all-small-caps","color": "#e85e56"},
+                children="QUERY"),
+            html.Div(id="current_query",style={"margin-left": "50px", "color": "#f4e9dc","width": "42%"},
+                children=""),
+            html.Div(style={"display":"flex"},children=[
+                html.Div(style={"width": "430px"},
+                    children=[html.Div(style={"margin-left": "30px", "font-variant": "all-small-caps","color": "#e85e56"},
+                                            children="DISPLAYED COLUMNS"),
+                                    dcc.Dropdown(id="r_table_paras",className="table-dropdown",style={"margin-left":"10%"},
+                                                options=opt,value=display_fileds,multi=True)]
+                        ),
+                html.Div(children=[html.Div(style={"font-variant": "all-small-caps","color": "#e85e56"},
+                                            children="SENTS PER PAGE"),
+                                    daq.NumericInput(id="n_pp",style={"margin-top": "4px","margin-left": "10%","color":"#f4e9dc","width": "40%"},
+                                            className="year-selector sent-pp",min=10,max=1000,value=16)]
+                        )
+                ]),
+            
+            
             html.Div(
-                id="rel_sent",style={"margin-left": "20px", "margin-top": "2%","font-variant": "all-small-caps"},className="side-by-side",
+                id="rel_sent",style={"position": "absolute","right": "1.5%","top": "90px","font-variant": "all-small-caps"},className="side-by-side",
                 children=[""]
             ),
             html.Div(
-                id="ranking-table",
+                id="ranking-table",style={"margin-top":"1%"},
                 className="output-datatable"
             ),
             ]),
@@ -554,106 +584,151 @@ def build_tab_1():
         )
     ]
 
-# --- build "sentences" tab ---
-# search, and show ranking
 
 @app.callback([Output("ranking-table", "children"),Output('memory-output', 'data'),
-                Output("rel_sent","children")],
+                Output("rel_sent","children"),Output("current_query","children")],
[…]
