### 2021-06-23 alicia-ziying-yang/conTEXT-explorer#1 @sara-02
[Suggestions on Improving JOSS Paper]
The current write-up is well put. I am adding some restricting suggestions to the content in the paper-based on JOSS's review requirement.

- [x] The `Statement of Need` section can be broken into 3 parts. The part up to page 2 para 1 `...analysis without programming knowledge.` forms the `statement of need` section.
- [x] The second para starting with `ConTEXT Explorer is developed using Dash..` and the 4 points after it can be moved under the `key features`.
- [x] The rest of the paragraphs forming the 3rd part can be renamed under a new section `Comparision with other tools` which basically covers JOSS's `Do the authors describe how this software compares to other commonly-used packages?` as a separate section.

While the content is all right as is, we can format the larger section into more useful categories.

JOSS Reference: https://github.com/openjournals/joss-reviews/issues/3347
