### 2021-06-23 #3347 @sara-02
@alicia-ziying-yang I have a few suggestions in the written material covered in this issue. https://github.com/alicia-ziying-yang/conTEXT-explorer/issues/1
