### 2021-07-19 alicia-ziying-yang/conTEXT-explorer#2 @faroit
[JOSS] Add automated unit tests
As part of the JOSS review, I was checking if the software has automated tests included which currently is not the case.

The JOSS guidelines state:

> Authors are strongly encouraged to include an automated test suite covering the core functionality of their software.

>     Good: An automated test suite hooked up to an external service such as Travis-CI or similar
>    OK: Documented manual steps that can be followed to objectively check the expected functionality of the software (e.g., a sample input file to assert behavior)
>     Bad (not acceptable): No way for you, the reviewer, to objectively assess whether the software works

https://github.com/openjournals/joss-reviews/issues/3347
