### 2021-07-20 alicia-ziying-yang/conTEXT-explorer#2 @alicia-ziying-yang
hi @faroit, we have a small example included in the software. If you (or users) install our software, and run it for the first time, you could see that there is an existing corpus called "sample_data" in the dropdown of the corpus, and a saved analysis "SSM test" in the analysis dropdown. Users could use it to explore the core functions of software.
We have mentioned this example in the README "How to use" section, after "Run the app".

If you have any further question, please let me know. Happy to solve.

Thank you.
