+++ doc/paper/paper.md
@@ -40,12 +40,7 @@ Similarly, there is a need for tools that assist in the manual construction of c
 
 Most existing computational methods underlying automated text processing require at least a working knowledge of relevant methods and programming languages (such as R or Python). `ConTEXT Explorer` is designed to lower these barriers to entry, by allowing an application of information retrieval and machine learning methods to text analysis without programming knowledge. 
 
-`ConTEXT Explorer` is developed using **Dash** [@plotly] in Python, and integrates the following packages.
-
-- **Spacy** pipeline [@spacy] - for pre-processing the text corpora uploaded by users.
-- **Whoosh** [@whoosh] - for building a search engine, which allows ranking of sentences relevant to the given concepts, and word frequency analysis at the sentence and document level.
-- **Gensim** [@gensim] - for training a word2vec [@word2vec] model for the uploaded corpus, which allows the user to find words related to other words for expanding concepts.
-- **Plotly** [@plotly] - for visualizing results in interactive graphs, which can be customized and saved as PNG files.
+# Comparision with other tools
 
 Current text analysis tools require either previous knowledge of programming (e.g., R, Python), or are commercial products (e.g., **RapidMiner** [@rapidminer], **Google Cloud Natural Language API** [@googlenlp]). One exception that we are aware of is **Voyant Tools** [@voyant], which is an open-source web-based text analysis tool built in Java. It allows the users to explore their data using some basic text analysis techniques such as word frequencies (at the document level), word cloud, and word context (words appearing around a chosen term). `ConTEXT Explorer` provides several functionalities that give a user a deeper understanding of the text, which are currently not available in Voyant Tools, such as: concept suggestions, sentence ranking and concept grouping. 
 It includes models allowing discovery of similar terms, and a search engine allowing retrieval of sentences relevant to concept terms, which can be used for concept expansion, and visualization of concepts over time. One key feature is that a caoncept can be either a a conjunction or disjunction of bags of words.
@@ -59,6 +54,13 @@ Compared to commercial text analysis systems such as **RapidMiner** [@rapidminer
  
 
 # Key features
+
+`ConTEXT Explorer` is developed using **Dash** [@plotly] in Python, and integrates the following packages.
+
+- **Spacy** pipeline [@spacy] - for pre-processing the text corpora uploaded by users.
+- **Whoosh** [@whoosh] - for building a search engine, which allows ranking of sentences relevant to the given concepts, and word frequency analysis at the sentence and document level.
+- **Gensim** [@gensim] - for training a word2vec [@word2vec] model for the uploaded corpus, which allows the user to find words related to other words for expanding concepts.
+- **Plotly** [@plotly] - for visualizing results in interactive graphs, which can be customized and saved as PNG files.
 
 `ConTEXT Explorer` has been tested locally under macos and as a server running under Ubuntu.
 
