### 2021-07-20 alicia-ziying-yang/conTEXT-explorer#1 @alicia-ziying-yang
Hi @sara-02 , I have reformat this section. Please check.
Thank you.
