+++ .travis.yml
@@ -1,0 +1,10 @@
+language: python
+python:
+  - "3.7.5"
+# command to install dependencies
+install:
+  - pip install -r requirements.txt
+  - python -m spacy download en
+# command to run tests
+script:
+  - python app.py
