### 2021-07-20 alicia-ziying-yang/conTEXT-explorer#2 @alicia-ziying-yang
Hi @faroit,
Have done with Travis-CI, and have attached the build-pass badge to the README.
Please check.

Thank you!
