### 2021-07-26 alicia-ziying-yang/conTEXT-explorer#2 @faroit
@alicia-ziying-yang thanks for adding this. For now, this only tests of the installation routine will pass and the software can be started. I suggest you also add some tests that test the functionality of the software and the data processing:

I guess https://dash.plotly.com/testing this would be a good start to read about testing with dash
