+++ .travis.yml
@@ -7,4 +7,12 @@ install:
   - python -m spacy download en
 # command to run tests
 script:
-  - nohup python app.py &
+  - python -m pytest ./
+
+before_script:
+    - wget http://chromedriver.storage.googleapis.com/2.27/chromedriver_linux64.zip
+    - unzip chromedriver_linux64.zip
+    - sudo apt-get install libnss3
+    - sudo apt-get --only-upgrade install google-chrome-stable
+    - sudo cp chromedriver /usr/local/bin/.
+    - sudo chmod +x /usr/local/bin/chromedriver
