+++ requirements.txt
@@ -14,4 +14,6 @@ gensim==3.8.1
 nltk==3.4.5
 backports.csv==1.0.7
 urllib3==1.25.8
-wordcloud==1.8.1 
+wordcloud==1.8.1
+pytest==5.3.5
+selenium==3.141.0
