+++ requirements.txt
@@ -17,3 +17,4 @@ urllib3==1.25.8
 wordcloud==1.8.1
 pytest==5.3.5
 selenium==3.141.0
+dash[testing]
