+++ requirements.txt
@@ -17,4 +17,3 @@ urllib3==1.25.8
 wordcloud==1.8.1
 pytest==5.3.5
 selenium==3.141.0
-dash[testing]
