### 2021-07-26 b7373e4
Update .travis.yml
