+++ .travis.yml
@@ -4,6 +4,7 @@ python:
 # command to install dependencies
 install:
   - pip install -r requirements.txt
+  - pip install dash[testing]
   - python -m spacy download en
 # command to run tests
 script:
