+++ .travis.yml
@@ -22,6 +22,14 @@ before_script:
     - wget http://chromedriver.storage.googleapis.com/2.27/chromedriver_linux64.zip
     - unzip chromedriver_linux64.zip
     - sudo apt-get install libnss3
+    - sudo apt-get install -y libglib2.0-0=2.50.3-2 \
+    libnss3=2:3.26.2-1.1+deb9u1 \
+    libgconf-2-4=3.2.6-4+b1 \
+    libfontconfig1=2.11.0-6.7+b1
     - sudo apt-get --only-upgrade install google-chrome-stable
     - sudo cp chromedriver /usr/local/bin/.
     - sudo chmod +x /usr/local/bin/chromedriver
+    - export DISPLAY=:99.0
+    - sh -e /etc/init.d/xvfb start
+    - sleep 3
+
