+++ .travis.yml
@@ -19,7 +19,7 @@ script:
   - python -m pytest ./
 
 before_script:
-    - wget http://chromedriver.storage.googleapis.com/2.27/chromedriver_linux64.zip
+    - wget https://chromedriver.storage.googleapis.com/92.0.4515.43/chromedriver_linux64.zip
     - unzip chromedriver_linux64.zip
     - sudo apt-get install libnss3
     - sudo apt-get install chromium-browser
