+++ .travis.yml
@@ -22,7 +22,6 @@ before_script:
     - wget https://chromedriver.storage.googleapis.com/92.0.4515.43/chromedriver_linux64.zip
     - unzip chromedriver_linux64.zip
     - sudo apt-get install libnss3
-    - sudo apt-get install chromium-browser
     - sudo apt-get --only-upgrade install google-chrome-stable
     - sudo cp chromedriver /usr/local/bin/.
     - sudo chmod +x /usr/local/bin/chromedriver
