+++ .travis.yml
@@ -22,6 +22,7 @@ before_script:
     - wget https://chromedriver.storage.googleapis.com/92.0.4515.43/chromedriver_linux64.zip
     - unzip chromedriver_linux64.zip
     - sudo apt-get install libnss3
-    - sudo apt-get --only-upgrade install google-chrome-stable
+    - wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
+    - dpkg -i google-chrome-stable_current_amd64.deb; apt-get -fy install
     - sudo cp chromedriver /usr/local/bin/.
     - sudo chmod +x /usr/local/bin/chromedriver
