+++ .travis.yml
@@ -26,3 +26,7 @@ before_script:
     - dpkg -i google-chrome-stable_current_amd64.deb; apt-get -fy install
     - sudo cp chromedriver /usr/local/bin/.
     - sudo chmod +x /usr/local/bin/chromedriver
+    - export DISPLAY=:99.0
+    - sh -e /etc/init.d/xvfb start
+    - sleep 3
+
