+++ .travis.yml
@@ -23,7 +23,7 @@ before_script:
     - unzip chromedriver_linux64.zip
     - sudo apt-get install libnss3
     - wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
-    - dpkg -i google-chrome-stable_current_amd64.deb; apt-get -fy install
+    - dpkg -i google-chrome-stable_current_amd64.deb; sudo apt-get -fy install
     - sudo cp chromedriver /usr/local/bin/.
     - sudo chmod +x /usr/local/bin/chromedriver
     - export DISPLAY=:99.0
