+++ .travis.yml
@@ -2,7 +2,8 @@ language: python
 python:
   - "3.7.5"
 # command to install dependencies
-
+services:
+  - xvfb
 addons:
   apt:
     sources:
@@ -27,6 +28,5 @@ before_script:
     - sudo cp chromedriver /usr/local/bin/.
     - sudo chmod +x /usr/local/bin/chromedriver
     - export DISPLAY=:99.0
-    - sh -e /etc/init.d/xvfb start
     - sleep 3
 
