+++ test_cases.py
@@ -1,0 +1,319 @@
+from dash.testing.application_runners import import_app
+import time
+
+# This first test case will test the main functionality of the app.
+# It will use the pre-install sample corpus.
+def test_main(dash_duo):
+	app = import_app("app")
+	dash_duo.start_server(app)
+	dash_duo.wait_for_text_to_equal("#corpus-select-title", "Use an existing corpus", timeout=4)
+	
+	#---test to choose the pre-installed sample data---
+	# choose corpus "sample_data"
+	corpus = dash_duo.driver.find_element_by_id('index-choose_corpus')
+	dash_duo.multiple_click(corpus, 1)
+	opt = dash_duo.driver.find_element_by_id('react-select-2--list')
+	dash_duo.multiple_click(opt, 1)
+	dash_duo.wait_for_text_to_equal("#react-select-2--value-item", "sample_data", timeout=3)
+
+	# choose to create a new analysis
+	saved_analysis = dash_duo.driver.find_element_by_id('index-choose-save')
+	dash_duo.wait_for_text_to_equal("#react-select-3--value-item", "create new", timeout=3)
+
+	#---start---
+	start = dash_duo.driver.find_element_by_id("start-exp")
+	dash_duo.multiple_click(start, 1)
+
+	# leave some time to load
+	time.sleep(2)
+
+	#---test year range---
+	year_from = dash_duo.find_element("#year-from > Div > span > input")
+	year_to = dash_duo.find_element("#year-to > Div > span > input")
+	# default values
+	assert '2012' == year_from.get_attribute('value')
+	assert '2018' == year_to.get_attribute('value')
+
+	# test to change value
+	dash_duo.clear_input(year_from)
+	year_from.send_keys('2013')
+	assert '2013' == year_from.get_attribute('value')
+
+	
+	#---test to type base term---
+	base_term = dash_duo.driver.find_element_by_id("base-term")
+	base_term.send_keys('marriage')
+	time.sleep(1)
+	assert 'marriage' == base_term.get_attribute('value')
+
+
+	#---test "Add Related Words"---
+	
+	# The word2vec in the app should find a list of 
+	# related words for the base term.
+	# In our test example, it should be the word "sex".
+	first_candidate = dash_duo.find_element("#candidates > div > div > label")
+	assert "sex" == first_candidate.text
+	dash_duo.multiple_click(first_candidate, 1)
+
+	# add it to the query
+	add_single = dash_duo.driver.find_element_by_id("add-single")
+	dash_duo.multiple_click(add_single, 1)
+
+	# delete it from the quey
+	added = dash_duo.find_element("#added-terms-fromside > div > div > label")
+	assert "sex" == added.text
+
+	dash_duo.multiple_click(added, 1)
+
+	# mannually add
+	manu_add = dash_duo.driver.find_element_by_id("manu-add")
+	manu_input = dash_duo.driver.find_element_by_id("manu-term")
+	manu_input.send_keys('equality') # single term
+	dash_duo.multiple_click(manu_add, 1)
+	dash_duo.clear_input(manu_input)
+	manu_input.send_keys('same sex') # phrase
+	dash_duo.multiple_click(manu_add, 1)
+
+
+	#---test "Sentence" tab---
+	sent_tab = dash_duo.driver.find_element_by_id("Specs-tab")
+	dash_duo.multiple_click(sent_tab, 1)
+	time.sleep(1)
+
+	# check the query for the ranking
+	current_query = dash_duo.driver.find_element_by_id("current_query")
[…]
