### 2021-07-27 alicia-ziying-yang/conTEXT-explorer#2 @alicia-ziying-yang
Hi @faroit,

Thanks for pointing out. Now I have added some test cases in https://github.com/alicia-ziying-yang/conTEXT-explorer/blob/main/test_cases.py
and change the Travis-CI .yml file.

Please check the newest build: https://travis-ci.com/github/alicia-ziying-yang/conTEXT-explorer

Thank you.
