### 2021-08-15 alicia-ziying-yang/conTEXT-explorer#3 @baileythegreen
Suggestions for ConTEXT-explorer dashboard
Here are some thoughts I've had when playing around with the dashboard. These are mostly UI/UX-related, and none of them are about things that are 'broken' or absolutely have to be fixed; they're just things I think would constitute improvements. I've put them all in one issue with headings so it's easier to read through, but am happy to resubmit some or all of them as individual issues so the suggestions can be considered on an individual basis.

## General
- Having all of the scroll bars remain visible all of the time would be an improvement. Right now, some of them seem permanently visible, but others disappear when the mouse isn't hovering over them (Overview–Top 30 Words; Groups–top section; Graphs–bottom section). The user could miss that these sections scroll, and not find the data they are looking for.
- The ability to download a report of some kind that is easily shareable and comparable between analyses would be a nice addition. Because the dashboard is interactive, maybe letting the user set things they way they want, then download the current configuration of the data would make sense.
- A 'save all' button for the graphs in the Groups and Graphs tab would similarly be very nice.
- Some way to share settings—like, when I save an analysis, does this produce a configuration file I could send to someone?

## Sentences
- This tab seems to reset automatically if you navigate away (to another tab), and then back. This may be frustrating for users.

## Graphs
- It might be worth considering whether users are going to use the Graphs tab to try to directly compare individual words/phrases from different sets of words (i.e., the contents of two of the bottom graphs). With the proportions, it may make sense to fix the y-axis to always show the full range, as this is fixed, so that direct comparison is possible.
- The graphs on this tab all have a 'Box select' and a 'Lasso select' option (the Top 30 Words one does, as well), that let users select a portion of the graph, but it doesn't appear that this has any bearing on anything else. I can't, for example, select part of a graph and download only that part, or zoom in on that part. It's unclear why the options exist.

## Save
- State on the boxes for each group whether they are an 'Any' or an 'All' grouping. The colour system is nice to reinforce this, but requires the user to remember the settings, and is a bit confusing because the colours are described as 'pink' and 'red' in the pdf, but against the dark background it's not at all clear that the light colour is pink, and the red colour is kind of a dark orangey-pink to my eye.

JOSS Reference: openjournals/joss-reviews#3347
