### 2021-08-15 alicia-ziying-yang/conTEXT-explorer#4 @baileythegreen
Suggestions for ConTEXT-explorer paper
I agree with the suggestion from #1 about moving the Key Features heading to before line 40 and renaming the third part.

## Typos
- Lines 15, 25: corpora is misspelled
- Line 19: original is misspelled
- Line 26: usually is misspelled
- Line 28: interested is misspelled
- Line 29, footnote 1: multiple is misspelled
- Line 63: concept is misspelled
- Figure 4: missing period at end of the caption

## Other things that need fixing
- Line 117: references Section 3.2, but your sections are not numbered

JOSS Reference: openjournals/joss-reviews#3347
