### 2021-08-15 alicia-ziying-yang/conTEXT-explorer#6 @baileythegreen
Analysis is case-sensitive
Analyses are case-sensitive and this is not made clear in the README or in the paper. I attach a screenshot demonstrating this with the first five chapters of Pride & Prejudice and the keywords: Elizabeth, Bennet, Fitzwilliam, and Darcy—both capitalised and uncapitalised. You can see the results for the Uppercase set do not match those for the Lowercase set.

I consider this something that needs to be—at a minimum—clarified before the paper is published. Really, I think this aspect of the behaviour should be altered so that it is not case-sensitive; but perhaps there is a good reason to keep it—with clarification.

JOSS Reference: openjournals/joss-reviews#3347
