### 2021-08-15 alicia-ziying-yang/conTEXT-explorer#7 @baileythegreen
Be more specific about target audience
Mention the target audience more specifically in the Readme, to meet this criteria:

`A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?`

JOSS Reference: openjournals/joss-reviews#3347
