### 2021-08-15 alicia-ziying-yang/conTEXT-explorer#8 @baileythegreen
Default is to run locally, but the Readme indicates otherwise
The Readme indicates that the Ubuntu server is the default option and action needs to be taken to run this locally; however, `app.py` is set to run locally by default.

## Readme
> If you want to run ConTEXT Explorer on your local computer, comment the code for ubuntu server, and uncomment the last line in app.py:
>
>  ```python
> # app.run_server(debug=False, host="0.0.0.0") # ubuntu server
> app.run_server(debug=False, port="8010") # local test
> ```
> You can then access the app through your browser: http://127.0.0.1:8010

### app.py
> ```python
> #app.run_server(debug=False,host="0.0.0.0") # ubuntu server
> app.run_server(debug=False, port="8010") # local test
> ```

JOSS Reference: openjournals/joss-reviews#3347
