+++ doc/paper/paper.md
@@ -31,11 +31,11 @@ bibliography: paper.bib
 
 # Statement of need
 
-With the availability of digital sources of data and associated tools, automated text analysis is becoming increasingly popular in humanities and social sciences. While for very large coprora, unsupervised text mining methods like topic modelling [@lda] can provide some useful summaries of data, many social science and humanities applications require analysis of data in context. That is, simple "bags of words" automatically mined and presented in isolation from the orignial text are often not meaningful for complex questions involving human behaviour and society. Inevitably, human interpretation is required to make sense of such patterns. For corpora with more than several hundred documents, there is a need for computational tools that can assist researchers in exploring the context in which "bags of words" (we will call them <em>concepts</em> from now on) occur. 
+With the availability of digital sources of data and associated tools, automated text analysis is becoming increasingly popular in humanities and social sciences. While for very large copora, unsupervised text mining methods like topic modelling [@lda] can provide some useful summaries of data, many social science and humanities applications require analysis of data in context. That is, simple "bags of words" automatically mined and presented in isolation from the original text are often not meaningful for complex questions involving human behaviour and society. Inevitably, human interpretation is required to make sense of such patterns. For corpora with more than several hundred documents, there is a need for computational tools that can assist researchers in exploring the context in which "bags of words" (we will call them <em>concepts</em> from now on) occur. 
 
-Similarly, there is a need for tools that assist in the manual construction of concepts from text coropora. In particular, manual intervention to judge the semantic intent of words (e.g., word sense disambiguation) is usualy needed to filter keywords to add to concepts that might be generated by automatic methods such as query expansion [@QE] or comparison of word embeddings [@word2vec]. For example, if a researcher is interesed in finding articles about same sex marriage, they might start with "same_sex marriage"[^1] as a concept. Automated methods processing a corpus of news articles might suggest related words like 'matrimony', 'union', 'erosion', and 'puzzlement'[^2]. Depending on the research question and the context of these words, some might be relevant to the concept and should be included, while others are not. It requires complex human judgement to make the distinction. `ConTEXT Explorer` is a tool to assist construction of such concepts in context.
+Similarly, there is a need for tools that assist in the manual construction of concepts from text coropora. In particular, manual intervention to judge the semantic intent of words (e.g., word sense disambiguation) is usually needed to filter keywords to add to concepts that might be generated by automatic methods such as query expansion [@QE] or comparison of word embeddings [@word2vec]. For example, if a researcher is interested in finding articles about same sex marriage, they might start with "same_sex marriage"[^1] as a concept. Automated methods processing a corpus of news articles might suggest related words like 'matrimony', 'union', 'erosion', and 'puzzlement'[^2]. Depending on the research question and the context of these words, some might be relevant to the concept and should be included, while others are not. It requires complex human judgement to make the distinction. `ConTEXT Explorer` is a tool to assist construction of such concepts in context.
 
-[^1]: We use underscore to join mulitple words into a single phrase.
+[^1]: We use underscore to join multiple words into a single phrase.
 [^2]: This is an example where we apply `ConTEXT Explorer` in the Australian Research Council Discovery Project (DP180101711) "Understanding Political Debate and Policy Decisions Using Big Data".
 
 Most existing computational methods underlying automated text processing require at least a working knowledge of relevant methods and programming languages (such as R or Python). `ConTEXT Explorer` is designed to lower these barriers to entry, by allowing an application of information retrieval and machine learning methods to text analysis without programming knowledge. 
[…]
