### 2021-08-16 alicia-ziying-yang/conTEXT-explorer#4 @alicia-ziying-yang
Hi @baileythegreen , thank you so much for pointing them out.
I have fixed the issues you mentioned, and pushed to the repo (paper.md).
