### 2021-08-16 alicia-ziying-yang/conTEXT-explorer#6 @alicia-ziying-yang
Hi @baileythegreen ,
Thank you so much for finding this issue (actually it's a bug).
I have fixed that, so now for example "marriage" and "Marriage" should be equal in the search process. That is, it is not case-sensitive search anymore.
