### 2021-08-18 alicia-ziying-yang/conTEXT-explorer#3 @alicia-ziying-yang
Hi @baileythegreen ,
Thank you so much for giving these ideas. Given the limited time and funding, we cannot implemented all of them now. So I listed some of them into the issues of the wish list. And for some others, please see the reply below.

**General**

- The ability to download a report of some kind that is easily shareable and comparable between analyses would be a nice addition. Because the dashboard is interactive, maybe letting the user set things they way they want, then download the current configuration of the data would make sense.   #9

- A 'save all' button for the graphs in the Groups and Graphs tab would similarly be very nice. #10

- Some way to share settings—like, when I save an analysis, does this produce a configuration file I could send to someone? (Re: the saved analysis will be "saved" into the application folders, and other users can reload all the "settings")

**Sentences**
This tab seems to reset automatically if you navigate away (to another tab), and then back. This may be frustrating for users.(Re: yes the sentences search the corpus again every time, because the query could be changed by users when users view other tabs.)

**Graphs**
- It might be worth considering whether users are going to use the Graphs tab to try to directly compare individual words/phrases from different sets of words (i.e., the contents of two of the bottom graphs). With the proportions, it may make sense to fix the y-axis to always show the full range, as this is fixed, so that direct comparison is possible. #11

- The graphs on this tab all have a 'Box select' and a 'Lasso select' option (the Top 30 Words one does, as well), that let users select a portion of the graph, but it doesn't appear that this has any bearing on anything else. I can't, for example, select part of a graph and download only that part, or zoom in on that part. It's unclear why the options exist. (Re: this is a issue for Plotly, we cannot change. Maybe users can simply make a screenshot for the zoomed graph.)

**Save**

- State on the boxes for each group whether they are an 'Any' or an 'All' grouping. The colour system is nice to reinforce this, but requires the user to remember the settings, and is a bit confusing because the colours are described as 'pink' and 'red' in the pdf, but against the dark background it's not at all clear that the light colour is pink, and the red colour is kind of a dark orangey-pink to my eye. #12
