### 2021-08-18 #3347 @alicia-ziying-yang
Hi @faroit ,

We have changed (the paper.md and the software) what we can do now for the great comments given by @baileythegreen and @sara-02 . For those UI ideas for the software, we have added them to the wish list in the issues.

Cheers,
Alicia
