### 2021-08-18 alicia-ziying-yang/conTEXT-explorer#3 @baileythegreen
Thanks for considering these!

I wanted to clarify that this suggestion:

> Some way to share settings—like, when I save an analysis, does this produce a configuration file I could send to someone?

is about sharing things with people who **do not** have access to the same computer. For example, I, like many people these days, work remotely, and have never even met the people in my team *in person*. Even if we were in the same place, we all have our own computers we would run analyses on. So, this:

> (Re: the saved analysis will be "saved" into the application folders, and other users can reload all the "settings")

doesn't address the point of the suggestion, and is why I tried to clarify by saying I wanted to 'send' it.
