### 2021-08-18 alicia-ziying-yang/conTEXT-explorer#3 @baileythegreen
Okay, then maybe add something to the Readme that explains where to find the config files. I have looked for something like that, but have not yet managed to find where they are stored; this may just be that I don't know how they are named, or what file extension they would use, but if that's the case, it won't be any different for the users.
