### 2021-08-23 #3347 @sara-02
@faroit I am 99% done with my checklist, one item remains and that is around the functionality for APIs.
In the source code files I do not see any doc-strings for the main functions and files. Additionally, I observed that pip8 standards are not uniformly followed especially while assigning/updating variables.
Is there any other place where the API level functionalities are listed?
