### 2021-08-23 #3347 @sara-02
Additionally, while I do see a `test_cases.py`, there should a line in the README about being able to run this file, and if any more packages need to be installed for testing well.
It will be helpful for future contributors.
Also, just food for thought, as yours is a UI-based system are you planning on adding any UI tests as well?
