### 2021-08-24 #3347 @alicia-ziying-yang
> @faroit I am 99% done with my checklist, one item remains and that is around the functionality for APIs.
> In the source code files I do not see any doc-strings for the main functions and files. Additionally, I observed that pip8 standards are not uniformly followed especially while assigning/updating variables.
> Please let me know what is the best way of judging the functionality checklist
> @alicia-ziying-yang can you point to some doc that I might have missed regarding this. Under the doc section in the repo, I found the manual for navigating the tool at a UI level, but nothing about APIs.

Hi @sara-02 ,
The user manual is for users but not programmers, thus there is no coding document there.

For the documents that explain what each function does are commented in the code body. But most of them are callback functions of Plotly Dash, and they are not supposed be re-used/called by additional functions. And the parameters of those callback functions are inputs/actions of users on the interface, they are kind of "fixed" and cannot be changed. So, in general, I just commented what each function does, but not expect them to be re-used or changed (which will surely break the callback logics).

In general, this app does not provide functions that could be re-used. If contributors want to add more functions, they need create new functions and extend UI components in app.py. For key parameters, I have added comments to the code so that contributors will know how to interact with them.
