### 2021-09-25 #3347 @baileythegreen
@faroit I have tried verifying the new functionality, but find I am now unable to run the software with my own data set. I tested it in a new virtual environment, with dependencies installed from the repo's `requirements.txt` file, and it still did not work. I'm not familiar enough with the inner workings to tell if this issue is caused by something on my end, or the software, but have given the traceback in an issue in the repo.
