### 2021-09-27 alicia-ziying-yang/conTEXT-explorer#14 @alicia-ziying-yang
Hi @baileythegreen ,
The problem may be due to your python version. Please refer : https://github.com/alicia-ziying-yang/conTEXT-explorer#install-required-dependencies

The python version has to be **3.7.5**.
And you may need to do: `python -m spacy download en` after installing all packages in the requirements.txt.
Could you please try these two, and see if there is still any error?

Thank you for checking!
