### 2021-09-27 #3347 @alicia-ziying-yang
> > @faroit I have tried verifying the new functionality, but find I am now unable to run the software with my own data set. I tested it in a new virtual environment, with dependencies installed from the repo's `requirements.txt` file, and it still did not work. I'm not familiar enough with the inner workings to tell if this issue is caused by something on my end, or the software, but have given the traceback in an issue in the repo.
>
> @baileythegreen thanks for letting me know. I will try to also look into this myself if this is useful. Is checking the updated functionalities the only remaining issue?

I put my reply here: https://github.com/alicia-ziying-yang/conTEXT-explorer/issues/14#issuecomment-927411163
Please check if they can fix the problem. Cheers:)
