### 2021-09-28 alicia-ziying-yang/conTEXT-explorer#15 @faroit
Add setup.py and publish package on pypi
Given that the software has certain specific python requirements, I suggest to package the software as a proper python package and define the python requirements there.

This issue is part of a JOSS review: https://github.com/openjournals/joss-reviews/issues/3347
