### 2021-09-28 alicia-ziying-yang/conTEXT-explorer#14 @faroit
@alicia-ziying-yang @baileythegreen i was able to install and run the software using python==3.7.5. Given similar issues could be avoided using python packaging, I opened #15.
