### 2021-09-29 99650fd
removed python setup.py build
