+++ setup.py
@@ -11,6 +11,7 @@ with open('requirements.txt', 'r') as f:
   install_requires = f.readlines()
 
 CUSTOM_COMMANDS = [
+  ['pip', 'install', 'spacy==2.2.4'],
   ['python', '-m', 'spacy', 'download', 'en'],
 ]
 
