### 2021-09-29 alicia-ziying-yang/conTEXT-explorer#15 @alicia-ziying-yang
Hi @faroit , I have created a setup.py file in the root, and update the installation info in README.md.
But we are not sure if we want to publish the current version to pypi. Thus I create a todo issue #16 .
Thanks for your suggestion!
