### 2021-09-29 #3347 @alicia-ziying-yang
Hi @faroit @baileythegreen ,
I have created a `setup.py` file in the root folder, and update the installation info in `README.md`.
If the python version is not 3.7.5, the installation will stop, and show error.
