### 2021-09-29 alicia-ziying-yang/conTEXT-explorer#17 @faroit
Relax python requirement
even though the installation issue is fixed in #15 I would suggest not tp freeze python to `python_requires='==3.7.5',` but instead find out the supported range of python versions to identify the problematic dependencies (and in turns fix the code support newer dependencies later)

https://github.com/openjournals/joss-reviews/issues/3347
