### 2021-09-29 alicia-ziying-yang/conTEXT-explorer#18 @faroit
Add CLI interface
When deploying to pypi (#16) running the default command `python app.py` wouldn't work obviously. It would therefore be beneficial if a real cli interface could be added.

This can be easily added into the setup.py as described [here](https://flask.palletsprojects.com/en/2.0.x/cli/#custom-scripts)

https://github.com/openjournals/joss-reviews/issues/3347
