### 2021-09-29 436d626
change with setup.py for fixing #18
