### 2021-09-29 alicia-ziying-yang/conTEXT-explorer#17 @faroit
> as we used a lot of packages that has version clashes with each other... for now, can only be put into wish list.

in that case i would suggest to add a conda environment file and a docker installation method
