### 2021-09-30 alicia-ziying-yang/conTEXT-explorer#17 @alicia-ziying-yang
I do not use docker, but can provide conda environment file: https://github.com/alicia-ziying-yang/conTEXT-explorer/blob/main/ce-env.yml
