### 2021-09-30 531bb8e
the conda environment file for this app
