+++ ce-env.yml
@@ -1,0 +1,100 @@
+name: ce-env
+channels:
+  - defaults
+dependencies:
+  - ca-certificates=2021.7.5=hecd8cb5_1
+  - certifi=2021.5.30=py37hecd8cb5_0
+  - libcxx=12.0.0=h2f01273_0
+  - libedit=3.1.20210714=h9ed2024_0
+  - libffi=3.2.1=h0a44026_1007
+  - ncurses=6.2=h0a44026_1
+  - openssl=1.1.1l=h9ed2024_0
+  - python=3.7.5=h359304d_0
+  - readline=7.0=h1de35cc_5
+  - sqlite=3.33.0=hffcf06c_0
+  - tk=8.6.11=h7bc2e8c_0
+  - wheel=0.37.0=pyhd3eb1b0_1
+  - xz=5.2.5=h1de35cc_0
+  - zlib=1.2.11=h1de35cc_3
+  - pip:
+    - attrs==21.2.0
+    - backports-csv==1.0.7
+    - bleach==4.1.0
+    - blis==0.4.1
+    - brotli==1.0.9
+    - build==0.7.0
+    - catalogue==1.0.0
+    - charset-normalizer==2.0.6
+    - click==8.0.1
+    - colorama==0.4.4
+    - context-explorer==1.0.0
+    - cycler==0.10.0
+    - cymem==2.0.5
+    - dash==1.14.0
+    - dash-bootstrap-components==0.10.5
+    - dash-core-components==1.10.2
+    - dash-daq==0.5.0
+    - dash-html-components==1.0.3
+    - dash-renderer==1.6.0
+    - dash-table==4.9.0
+    - dash-uploader==0.4.1
+    - docutils==0.17.1
+    - en-core-web-sm==2.2.5
+    - flask==2.0.1
+    - flask-compress==1.10.1
+    - future==0.18.2
+    - gensim==3.8.1
+    - gunicorn==20.1.0
+    - idna==3.2
+    - importlib-metadata==4.8.1
+    - itsdangerous==2.0.1
+    - jinja2==3.0.1
+    - keyring==23.2.1
+    - kiwisolver==1.3.2
+    - markupsafe==2.0.1
+    - matplotlib==3.4.3
+    - more-itertools==8.10.0
+    - murmurhash==1.0.5
+    - nltk==3.4.5
+    - numpy==1.21.2
+    - packaging==21.0
+    - pandas==1.2.3
+    - pep517==0.11.0
+    - pillow==8.3.2
+    - pip==21.2.4
+    - pkginfo==1.7.1
+    - plac==1.1.3
+    - plotly==5.3.1
+    - pluggy==0.13.1
+    - preshed==3.0.5
+    - py==1.10.0
+    - pygments==2.10.0
+    - pyparsing==2.4.7
+    - pytest==5.3.5
+    - python-dateutil==2.8.2
+    - pytz==2021.1
+    - readme-renderer==29.0
+    - requests==2.26.0
+    - requests-toolbelt==0.9.1
+    - rfc3986==1.5.0
+    - scipy==1.7.1
+    - selenium==3.141.0
+    - setuptools==58.1.0
+    - six==1.16.0
+    - smart-open==5.2.1
+    - spacy==2.2.4
+    - srsly==1.0.5
+    - tenacity==8.0.1
+    - thinc==7.4.0
+    - tomli==1.2.1
+    - tqdm==4.62.3
+    - twine==3.4.2
+    - typing-extensions==3.10.0.2
+    - urllib3==1.25.8
+    - wasabi==0.8.2
+    - wcwidth==0.2.5
+    - webencodings==0.5.1
+    - werkzeug==2.0.1
+    - whoosh==2.7.4
+    - wordcloud==1.8.1
+    - zipp==3.5.1
