### 2021-09-30 alicia-ziying-yang/conTEXT-explorer#17 @faroit
> I do not use docker, but can provide conda environment file: https://github.com/alicia-ziying-yang/conTEXT-explorer/blob/main/ce-env.yml

just tried this, please also add `pip` to the conda dependencies

```
Warning: you have pip-installed dependencies in your environment file, but you do not list pip itself as one of your conda dependencies.  Conda may not use the correct pip to install your packages, and they may end up in the wrong place.  Please add an explicit pip dependency.  I'm adding one for you, but still nagging you.
```
