+++ ce-env.yml
@@ -10,6 +10,7 @@ dependencies:
   - ncurses=6.2=h0a44026_1
   - openssl=1.1.1l=h9ed2024_0
   - python=3.7.5=h359304d_0
+  - pip
   - readline=7.0=h1de35cc_5
   - sqlite=3.33.0=hffcf06c_0
   - tk=8.6.11=h7bc2e8c_0
@@ -59,7 +60,6 @@ dependencies:
     - pandas==1.2.3
     - pep517==0.11.0
     - pillow==8.3.2
-    - pip==21.2.4
     - pkginfo==1.7.1
     - plac==1.1.3
     - plotly==5.3.1
@@ -95,4 +95,5 @@ dependencies:
     - werkzeug==2.0.1
     - whoosh==2.7.4
     - wordcloud==1.8.1
+    - https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz
     - zipp==3.5.1
