### 2021-09-30 fa7879d
update conda env setting
