### 2021-09-30 4599738
update the conda env setting
