### 2021-11-16 alicia-ziying-yang/conTEXT-explorer#19 @alicia-ziying-yang
Hi @baileythegreen ,

I used your .csv, and it worked for me well without any error.
But I am trying to figure out why you got those errors. Could you please check:

I saw you got `length of data: 0 ; length of corpus 0` in your log, which means that the app read nothing. But in my log, I got `length of data: 5 ; length of corpus 5` (using your dataset). The runtime error actually says the same thing: it has nothing to build vocabulary.

May I know what you set in Step 2 when uploading? What did you choose for the "sentences"? This is my setting:

Could you please try again using this setting?
