### 2021-11-16 alicia-ziying-yang/conTEXT-explorer#19 @alicia-ziying-yang
I have your dataset uploaded ok with the setting above. And in step 3, the name of the corpus is "pride_chapters". @baileythegreen
