### 2021-11-16 alicia-ziying-yang/conTEXT-explorer#19 @baileythegreen
Okay, I think I must not have been paying enough attention yesterday. I registered that the Year field was looking at the Year column, but the ID column was not looking at any columns. I was setting that one to Chapter, but apparently the Sentences field was escaping my notice.

Some kind of visible warning, error message, pop-up, et cetera in the dashboard might be helpful for users—for this, or other issues. I know to look at the terminal window, but your users might not.

Anyway, everything is working now and I feel a bit silly. The new installation process is very nice, though!
