### 2021-11-16 #3347 @baileythegreen
@faroit Apoligies for the delay(s), I've had many things to balance.

The new installation procedure is indeed, very nice, and I imagine it will pre-empt many problems. I have been able to upload new data and run queries, and obtain results that look as I would expect. I'm satisfied with this, and have completed my checklist.
