+++ .travis.yml
@@ -12,7 +12,7 @@ addons:
     - google-chrome-stable
     
 install:
-  - pip install --upgrade setuptools
+  - pip install --upgrade packaging
   - pip install -r requirements.txt
   - pip install dash[testing]
   - python -m spacy download en
