+++ .travis.yml
@@ -21,7 +21,7 @@ script:
   - python -m pytest ./
 
 before_script:
-    - wget https://chromedriver.storage.googleapis.com/94.0.4606.61/chromedriver_linux64.zip
+    - wget https://chromedriver.storage.googleapis.com/96.0.4664.93/chromedriver_linux64.zip
     - unzip chromedriver_linux64.zip
     - sudo apt-get install libnss3
     - wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
