### 2021-12-08 #3347 @danielskatz
👋 @alicia-ziying-yang - I've suggested a few small changes in the bib and md files in https://github.com/alicia-ziying-yang/conTEXT-explorer/pull/21 - Please merge this, and then be sure the change the in the md file is correct.  Let me know when you are then ready for final publication of this work.
